[#394] WidgetAnalyticsManager 이벤트 전송 로직을 메인 앱 전용 파일로 분리
